requirements.txt
# Core LangChain & LangGraph
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0

# MongoDB
pymongo>=4.6.0

# Web Scraping
firecrawl-py>=0.0.5
beautifulsoup4>=4.12.0  # LinkedIn job scraping (GAP-065)
lxml>=5.0.0  # HTML parser for BeautifulSoup
python-jobspy>=1.1.62  # Indeed job scraping for auto-ingestion

# Document Generation
python-docx>=1.1.0
pypandoc>=1.13
playwright>=1.40.0

# Google APIs
google-auth>=2.28.0
google-api-python-client>=2.120.0
gspread>=5.12.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
json-repair>=0.30.0  # Fix malformed LLM JSON outputs
importlib-metadata>=6.0.0; python_version < "3.10"  # Backport only for older runtimes
langfuse>=3.0.0  # Search observability for runnerless pipeline rollout
PyYAML>=6.0  # YAML-preferred source files for 4.3 master-CV / Codex authoring path

# Optional: Linting (encoding-rule guard for the cp1252 hazard class)
ruff>=0.7.0

# Optional: Testing
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-asyncio>=1.3.0
pytest-xdist>=3.5.0  # Parallel test execution (GAP-063)
pytest-cov>=4.0.0  # Coverage reporting
mongomock>=4.1.0  # In-memory MongoDB for unit tests (preenrich worker)
fakeredis>=2.20.0  # In-memory Redis for unit tests (preenrich outbox)
# pytest-playwright>=0.4.0  # For E2E tests - disabled until proper implementation
httpx>=0.28.0  # For FastAPI TestClient

# Runner service (FastAPI + ASGI server)
fastapi>=0.115.0
uvicorn[standard]>=0.24.0

# Frontend service (Flask)
Flask>=3.0.0
requests>=2.31.0
pydantic-settings>=2.0.0

# Queue management (Redis-backed)
redis>=5.0.0
flask-sock>=0.7.0
websocket-client>=1.6.0

# Annotation suggestion system - heavy ML deps in requirements-ml.txt
# sentence-transformers moved to requirements-ml.txt (pulls torch ~200MB+)
# Tests mock the embedding functionality, production installs both files
numpy>=1.24.0  # Used directly by annotation_priors.py, annotation_suggester.py
